write a different opt for metal
